Add files via upload
Developed a machine learning model for the WELFake dataset, a comprehensive collection of real and fake news articles. The dataset includes crucial elements such as Serial number, Title, Text, and Label (0 for fake, 1 for real), facilitating a detailed and systematic analysis of the content. Leveraging this diverse dataset, I implemented an effective machine learning algorithm to discern patterns and accurately classify articles, contributing to a nuanced understanding of news authenticity within the WELFake dataset.